+++ app/src/main/java/com/example/weatherapp3/weatherApi/Wind.java
@@ -10,7 +10,7 @@ public class Wind {
     private Double speed;
     @SerializedName("deg")
     @Expose
-    private Double deg;
+    private Integer deg;
 
     public Double getSpeed() {
         return speed;
@@ -20,11 +20,11 @@ public class Wind {
         this.speed = speed;
     }
 
-    public Double getDeg() {
+    public Integer getDeg() {
         return deg;
     }
 
-    public void setDeg(Double deg) {
+    public void setDeg(Integer deg) {
         this.deg = deg;
     }
